@@ -237,7 +237,7 @@
         </div>
     {/if}
     <!-- Rival -->
-    <div class="infoSlot infoRival" on:click={() => changeManager(viewManager.rival.link)}>
+    <div class="infoSlot infoRival" onclick={() => changeManager(viewManager.rival.link)}>
         <div class="infoLabel">
             Rival
         </div>
