@@ -481,7 +481,7 @@
 </style>
 
 <div class="matchup">
-    <div class="header" on:click={() => expandClose()} bind:this={el} >
+    <div class="header" onclick={() => expandClose()} bind:this={el} >
         <div class="opponent home{winning == "home" ? " homeGlow" : ""}">
             <img class="avatar" src={home.manager.avatar} alt="home team avatar" />
             <div class="name">{home.manager.name}</div>
@@ -554,7 +554,7 @@
             </div>
         {/each}
         {#if !expandOverride}
-            <div class="close" on:click={() => expandClose()}>Close Matchup</div>
+            <div class="close" onclick={() => expandClose()}>Close Matchup</div>
         {/if}
     </div>
 </div>
