@@ -1,27 +1,21 @@
 <script>
-	import { tabs } from '$lib/utils/tabs';
 	import NavSmall from './NavSmall.svelte';
 	import NavLarge from './NavLarge.svelte';
-    import { page } from '$app/stores';	
+    import { page } from '$app/state';
 	import IconButton from '@smui/icon-button';
 	import { Icon } from '@smui/common';
 
-	$: active = tabs.find(tab => tab.dest == $page.url.pathname || (tab.nest && tab.children.find(subTab => subTab.dest == $page.url.pathname)));
-
 	// toggle dark mode
-	let lightTheme =
-		typeof window === "undefined" ||
-		window.matchMedia("(prefers-color-scheme: light)").matches;
-	
-	function switchTheme() {
-		lightTheme = !lightTheme;
+	let darkTheme = $state(typeof window === "undefined" || window.matchMedia("(prefers-color-scheme: dark)").matches);
+	function switchTheme(currentTheme) {
+		currentTheme = !currentTheme;
 		let themeLink = document.head.querySelector("#theme");
 		if (!themeLink) {
 			themeLink = document.createElement("link");
 			themeLink.rel = "stylesheet";
 			themeLink.id = "theme";
 		}
-		themeLink.href = `/smui${lightTheme ? "" : "-dark"}.css`;
+		themeLink.href = `/smui${currentTheme ? "" : "-dark"}.css`;
 		document.head
 		.querySelector('link[href="/smui-dark.css"]')
 		.insertAdjacentElement("afterend", themeLink);
@@ -29,7 +23,7 @@
 </script>
 
 <svelte:head>
-	<title>{!$page.url.pathname[1] ? 'Home' : $page.url.pathname[1].toUpperCase() + $page.url.pathname.slice(2)} | League Page</title>
+	<title>{!page.url.pathname[1] ? 'Home' : page.url.pathname[1].toUpperCase() + page.url.pathname.slice(2)} | League Page</title>
 </svelte:head>
 
 <style>
@@ -87,8 +81,8 @@
 	<div class="container">
 		<IconButton
 			toggle
-			pressed={lightTheme}
-			on:MDCIconButtonToggle:change={switchTheme}
+			bind:pressed={darkTheme}
+			onclick={() => switchTheme(darkTheme)}
 			class="lightDark"
 		>
 			<Icon class="material-icons" on>dark_mode</Icon>
@@ -97,11 +91,11 @@
 	</div>
 
 	<div class="large">
-		<NavLarge {tabs} bind:active={active} />
+		<NavLarge />
 	</div>
 
 	<div class="small">
-		<NavSmall {tabs} bind:active={$page.url.pathname} />
+		<NavSmall />
 	</div>
 
 </nav>
